refect: 유틸리티 함수 용도별 분리
