+++ src/pages/itemsPage/ProductIdPage/components/ProductItem.jsx
@@ -1,19 +1,10 @@
 import user from "@/assets/imgs/user.png";
 import HeartSvg from "@/assets/svg/HeartSvg";
 import KebebSvg from "@/assets/svg/KebebSvg";
+import { formatDate } from "@/utils/date";
 import * as S from "./ProductItem.styles";
 
 const ProductItem = ({ item }) => {
-  function formatDate(isoString) {
-    const date = new Date(isoString);
-
-    const year = date.getFullYear();
-    const month = String(date.getMonth() + 1).padStart(2, "0");
-    const day = String(date.getDate()).padStart(2, "0");
-
-    return `${year}. ${month}. ${day}`;
-  }
-
   return (
     <S.ProductItemWrapper>
       <S.ProductItemImgWrapper>
