+++ src/utils/common.js
@@ -1,0 +1,2 @@
+/** 값의 최소 길이 검사 */
+export const hasMinLength = (v, min) => v.length > min;
